REAL to DOUBLE PRECISION

+++ PostgreSQL Skripts/CREATE.sql
@@ -31,8 +31,8 @@ CREATE TABLE employee (
     employee_id SERIAL PRIMARY KEY,
     first_name VARCHAR(200), 
     last_name VARCHAR(200),
-    base_fte REAL,
-    free_fte REAL,
+    base_fte DOUBLE PRECISION,
+    free_fte DOUBLE PRECISION,
     e_mail VARCHAR(200),
     phone_number VARCHAR(25), 
     entry_date DATE,
@@ -100,8 +100,8 @@ CREATE TABLE project (
     proj_name VARCHAR(35),
     proj_priority VARCHAR(15),
     proj_manager INT,
-    needed_fte REAL,
-    current_fte REAL,
+    needed_fte DOUBLE PRECISION,
+    current_fte DOUBLE PRECISION,
     start_date DATE,
     end_date DATE
     -- Foreign key constraints will be added after table creation
@@ -127,7 +127,7 @@ CREATE TABLE connection_team_employee (
     employee_id INT,
     FOREIGN KEY (team_id) REFERENCES team(team_id) ON DELETE CASCADE,
     FOREIGN KEY (employee_id) REFERENCES employee(employee_id) ON DELETE CASCADE,
-    assigned_fte REAL
+    assigned_fte DOUBLE PRECISION
 );
 
 CREATE TABLE connection_education_skill (
